feat: 运动积分支持小数+移动端榜单随机头像(业务积分带小数,全链路Long改BigDecimal:导入解析去掉整数限定改数字校验,默认保留两位四舍五入并去尾随0防1E+2科学计数法,38位精度预拦截成行级错误;手工编辑同样两位归一;榜单VO与全表累计积分统计BigDecimal化;前端编辑弹窗校验/占位/导入提示/下载模板样例12.5同步;移动端/points头像原显示手机号末位数字难看,改为按姓名散列确定性分配24种动物emoji+6色渐变底,同一人永远同一头像刷新不变,领奖台冠亚季军色与奖牌保留,人均积分改保留两位;建表存档points列注释更新,列本身NUMBER无需DDL;本地起服务验证rank接口JSON与移动端页面截图通过,推送后CI自动部署)

+++ project/src/main/java/com/bcsport/admin/entity/bcp/BcpSportPoints.java
@@ -4,6 +4,7 @@ import com.baomidou.mybatisplus.annotation.*;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -21,8 +22,8 @@ public class BcpSportPoints implements Serializable {
     /** 运动员/玩家名（导入去重键，未删除记录内唯一） */
     private String sporter;
 
-    /** 积分 */
-    private Long points;
+    /** 积分（NUMBER，支持小数） */
+    private BigDecimal points;
 
     @TableField(fill = FieldFill.INSERT)
     private LocalDateTime createTime;
