+++ project/src/main/java/com/bcsport/admin/vo/SportPointsRankVO.java
@@ -2,6 +2,8 @@ package com.bcsport.admin.vo;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * 运动积分排名视图对象（移动端免登录榜单页专用，只暴露展示必需字段）
  */
@@ -13,7 +15,7 @@ public class SportPointsRankVO {
     /** 运动员姓名（BC_SPORTS_BCP_SPORT_POINTS.sporter） */
     private String name;
 
-    private Long points;
+    private BigDecimal points;
 
     /** 全榜绝对名次（积分降序、同分按姓名稳定排序后的位次） */
     private Long rank;
